Refactored serial connection to be partially generic
